Make individual build pages match the theme.
In order to do this, I had to add a class (subpage) to the body
elements of those HTML files that get hosted in an IFrame
(which was much less work than adding the main class to the body
element of each build's page). Then swap the CSS selector logic
from body.main to body:not(.subppage).